Use OpenPGP instead of GPG when referring to the spec or its artifacts
While GPG is still one of the most prevalent implementations, it is just
one among many. We are dealing here with OpenPGP artifacts, so spell
that out explicitly. This is more relevant now that GnuPG has opted out
of the IETF OpenPGP working group, and will not be implementing future
revisions of the specification.